создана первая форма

+++ ServiceControl/Modbus/Devices/Device.cs
@@ -21,10 +21,10 @@ namespace ServiceControl.Modbus.Registers
         //----------------------------------------------------------------------------------------------
         // Конструктор
         //----------------------------------------------------------------------------------------------
-        public Device(MbWork modb, byte slave)
+        public Device(MbWork modb, int slave)
         {
             modbus = modb;
-            Slave = slave;
+            Slave = (byte)slave;
         }
